# audioverse-dotnet/AudioVerse.API/swagger.audioverse.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AudioVerse.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/admin/system-config": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSystemConfigurationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/login-attempts/{userId}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/login-attempts/recent-failed": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "minutes",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 15
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaptchaOption": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "UpdateSystemConfigurationRequest": {
        "type": "object",
        "properties": {
          "sessionTimeoutMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "captchaOption": {
            "$ref": "#/components/schemas/CaptchaOption"
          },
          "maxMicrophonePlayers": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}